+++ CHANGELOG.md
@@ -206,6 +206,7 @@ as in the new truchet animation examples, and examples/animations/ec.py)
 - Argument-less boolean operation methods for plural `P`
 - Automatic _-prefixed versions of all P methods, to get clojure-style "phrase"-commenting
 - `VERSIONS=` support for macro-like reversioning of a single animation
+- `Font.LibraryFind` and `Font.LibraryList` to search font registry (mac-only)
 ### Changed
 - `Mondrian` -> `Scaffold`
 - `th` & `tv` (true-horizontal & true-vertical) have been renamed to potential confusion (since `th` could be taken to mean `true-height`) — these are now `tx` and `ty` respectively